Use int32_t rather long on strategic places, to be more sparc64-friendly; ok jason@ a long time ago

+++ sys/crypto/sk.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: sk.h,v 1.2 1997/02/24 14:06:53 niklas Exp $	*/
+/*	$OpenBSD: sk.h,v 1.1 2000/02/28 23:13:05 deraadt Exp $	*/
 
 /* lib/des/sk.h */
 /* Copyright (C) 1995 Eric Young (eay@mincom.oz.au)
@@ -47,7 +47,7 @@
  * [including the GNU Public Licence.]
  */
 
-static const unsigned long des_skb[8][64]={
+static const u_int32_t des_skb[8][64]={
 {
 /* for C bits (numbered as per FIPS 46) 1 2 3 4 5 6 */
 0x00000000L,0x00000010L,0x20000000L,0x20000010L,
